don't longjmp (lua_error) from generated code
Add ->error function to MatchState. It changes scurrent to
special value and makes epilogue (which restores stack position
as it was before generated code). lpjit_match recognizes special
value and calls lua_error. Error message is already on stack.
All functions in lpjit_lpeg.c get mstate as first argument.
It is needed to call ->error function on it.